+++ cmd/consts.go
@@ -1,0 +1,4 @@
+package cmd
+
+const enpoint_openai = "https://api.openai.com/v1/"
+const endpoint_etsy = "https://openapi.etsy.com/v2/"
